Add find_requests service and requests API data
At the moment we don't do a great job when it comes to communicating
what's going on under the hood when you push a build to github - it
either shows on the builds list or not. In the latter case users may be
left in the dark, which I suspect may be more common in the future, when
we move such things like "exclude branch" to the repository settings
pane.

This commit reveals part of the request data, which we can use to show
what did we get from github and if it was accepted.